Adds sorting and filling within files.
Initial implementation. There is a lot of room for improvement.

I think one of the most imporant things to do is to turn the filling
routine into a callable after initialization, so files outside the
bounds can be skipped straight away and only files that will be used
in final aggregation will be sorted.

Right now, especially bad for running from cli, the day paddings will be
sorted. Would pseudo achieve the effect of clipping to the bounds of the
file since usually the file contents are mostly sorted.