Hook premade configurations into repo
At some point, we're going to switch to using these premade
configurations to be sure that aggregates are consistent for a product.
It would be awkward if a new variable showed up one day and started
getting included automatically in the aggregations. Other reasons
include being able to split products (eg seis east and west), possible
mag inboard and outboard, as well as configure chunking for variables
differently than in it is in the granular files.