bug fix: edge case fixed for signle file agg
When there was one file in the aggregation list, and the stop bound
was below the tolerance for timestep wiggle from the aggregation end
time, the else statement (aggregator.py#L215 was triggered but would
set the stop to be 0, so the aggregation would not actually have anything
in it.

Fixed this by skipping any setting of dim slice start or stop if num_overlap
is 0.