Allow UDC valid'n if test file has len 0 dims
If the first file in the list of files to aggregate is a valid netcdf of
the correct format, but has no data and the unlimited dimension has size
0, then validation used to fail.

This was undesireable because further files in the list could have data
and the aggregation list should have still been built.

Though we don't know if the other dim index is valid if dim size is 0
for the first file, we'll have to let this slide... a proper solution
would be to postpone validation of that configuration parameter until we
have a non-zero size dimension (ie. have data), not sure extra work is
worth the small convenience of so thouroughly validating that.

Fixes #7