Tests updated to use new configurations
Hmm, forgot to mention previously, another large change was
getting rid of the OO interface to the aggregator. There was
literally no reason to make the aggregator an object you had
to initialize.

This inflicted a pretty large change to the test suite. While
revising them to the new simply function based interface, many
of the tests were revised and improved to be more stringent.

Note: currently subsetting a dimension is not yet supported... ie
the SGPS East test fails currently.