Isolating general agg fn'ty as ncagg module
Summary: rename aggregoes -> ncagg for generatlity

ncagg has applications beyond NCEI and the GOES-R seires data, so
strategically isolating the core aggregation functionality into an
ncagg library and utility.

ncei specific wrappers and configuration do not need to be included
and should instead be a separate package that depends on this.