Add template-ncagg cmd for init custom templates
template-ncagg is a new command line interface that will print
the default configuration for a sample netcdf file given. The
config printed is the default used when no configuration template
is given.

By saving it to a file and then modifying it, many additional
aspects of the aggregation can be customized.

Also updates the documentation.