ver bump, allow spec of new dims through config
New dimensions specified in the config are implicitly a dimension
representing dependence on file and so use size 1 for each input file.

This feature has only been tested for default UDC configuration,
use with caution in cases of fill values, etc. (TODO: tests for that!)

Version bump and adds CHANGELOG!

Fixes #4.