Optimize imports, tentative fix for masked out of range vals
Optimize imports, simple code transform from Pycharm.

Tentative fix for masked values in incoming files. This has been
done so that in cases where the template changes, for example,
the valid_min, or the valid_max values, the values are not already
masked when read in, and therefor still masked when going out.