version bump, clean release
Previous release accidentally included some ustaged dev/debug changes,
so a version bump is necessary to release a new version to pip.

Updates the readme with a note about being careful not to build from
a repo with unstaged changes because they will end up in the build!

Maybe it's time to think about setting up a CI...