rm AggreList obj, 1 rec grans w/ overlaps, simp tests
AggreList object really wasn't doing anything useful. It
has been removed. Eg, despite trying to check class membership
of incoming objects, it was still possible to initialized
the AggreList with non AbstractNode subclasses. Just isn't worth
the effort to have type checking. It will break if used
incorrectly :P

We had a problem with EHIS files, two processing streams both
producing outputs that would be offset by 1 minute on this
5 minute product. Both streams would get included in the
output aggregation. Final cause: there was a fundamental problem
in how it was determining and fixing overlap.

Simplify testing. There were a lot of tests that were not
accurately testing the final outcome. Though not a comprehenzive
effort, this has been fixed in many of the tests.

Building aggregation is somewhat time costly, so in order to
make the test suite faster, in places where possible, the tests
were revised to only create one output aggregation and perform
all checks on that since output, instead of regenerating the same
thing many times, performing slightly different checks each time.

Also, I've tried to avoid testing the number of objects that
come out in an Aggregation List. Occasionally, though very rarely
this number should change. For many of these tests that are
evaluating the final output, the number of elements in the
intermediate list is an implemntation detail that shuldn't be
tied down.

A more appropraite place would be a number of tests specifcally
testing the creation of Aggregation Lists, but currently, we
don't have any like that.