Fix chopping first record if too close to bound
The first record was being chopped if it came at less
than 0.5*cadence after the boundary. Note that the boundary
is not real or valid data and so the first point should not
be chopped for being "too close but after it".

Some logic introduced to override gap_between to a pass
through value shoudl the first record be in danger of getting
chopped.

Added a test to check.

Furthermore, fixing this highlighted an issue with the SEIS
EHIS test.