Optimization: hold nc Dataset object in InputFileNode
Wow! Here's where the poerformance killer was. I guess this
wasn't immediately clear from the profile!

Previously: InputFileNode held a self.filename reference to
the filename string on disk and every time the netcdf was
needed, it was constructed.

Changed to:

Now: Open the netcdf in InputFileNode __init__ and refer to
that throughout.

--------------

Test suite timings are now ~28 and 29 seconds for 2.7 and 3.6,
down from 80 and 89 second spreviously.