sub dt_nom from start boundary instead of dt_min
This issue frequently resulted in chopping the first
record from the day since it would appear to fall too
closely to the boundary. Subtract instead dt_nom.

The result will be allowing (dt_nom - dt_min) more
wiggle room in the timing of the first record after
the start boundary.

I think this should now cover time. The important
check being that when aggregating a very long series
of files that all records end up in one aggregation
or another. This issue was previously chopping out
one record at the beginning of the day that would
neither appear in the last position of the previous
day nor in the first position of the next day.

Bumping version because effect of this change will
be relatively major for GOES-R datasets.