prevent index error when final is empty
In generate_aggregation_list, sometimes the preliminary list
has files, but none of them end up in the final.

This would result in an IndexError for this statement. Now check
for len(final) > 0 before trying to reference an element in final.

If len(final) == 0, this will fall through and return an empty
list, which is acceptable.

... if not isinstance(final[-1], FillNode):
IndexError: list index out of range