Fix comments about bound calculation
The comment that was left here about handling the bound
reflected the previous confusion I was in regarding this.
It seems to be worked out now. Originally, I was trying
to subtract something other than dt_min from prev_end in
order to correct num_overlap. It appears that prev_end
was correct all along and the real problem was with the
actual calculation of num_overlap.

I've reprocessed a number of files with 8.1 and am gaining
confidence that this has really been resolved.