Perfects filling calculation. Bump v 0.8.7
Perhaps a little soon after v0.8.6 to release
another version, but this is what I've been
working on for the last few days.

The recent issue with EXIS-L1b-XRSF files has
highlighted that there are still some issues
with the way gap filling worked. There were some
incorrect gaps both between files and inside files.

The major thing I've realized is that the when there
are previous files in the final aggregation list,
we need to round the size of the FillNode, otherwise
it needs to be floored.

I was having trouble with this calculation in the
past and it makes sense looking back that there was
trouble in trying to fit these different cases in
the same expression.

Another change... gap between must be a minimum size
and then if execution enters the conditional to add
a FillNode, we will definitely add a fill node by
taking the max(1, size). Similar in the InputNode
file_internal_aggregation_list.