Rm problem test file, adjust start fill calc
Fixes #3 (unit tests break with netCDF4 == 1.5.{2,3}) by removing
a problematic netcdf that was causing a double free error in the
NetCDF4-Python library. Not ultimately sure what was wrong with
this file, but moving on without it.

Furthermore, making a slight tweak to calculation of number of recs
necessary to fill at the start of a file by conditionally bumping the
value if necessary.