new copy_from_alt config param
New copy_from_alt config param for variables allows a template to
specify alternate variables from which the data will be copied from if
the "name" isn't specified.

So in summary: output data will be populated from:

1. First priority: the variable of same name from granule.
2. Alternatively: the first variable in copy_from_alt found in granule.

This change is intended to support a rename we are expecting to receive,
where we will also want to go back and put the old name into the new
variable as well.